V1.3
-基本满足需求
-如果日记已经存在, 可以追加内容
-查询时判断输入的文件名是否已经存在

+++ _src/om2py0w/0wex1/main.py
@@ -7,23 +7,42 @@ from os.path import exists
 
 script, text = argv
 
-txt = open(text, 'w')
+#判断输入的日记名称是否重复,如果输入的日记名称已经存在, 可以继续追加
+if exists(text) == True:
+	print """
+	日记文档已经存在!!!
+	是否要补充此日记? 输入 y继续 n退出
+	"""
+	judge1 = raw_input()
 
+	#确认补充日记
+	if judge1 == "y":
+		content1 = raw_input("补充日记:")
+		txt1 = open(text, 'a')
+		txt1.write(content1)
+		txt1.close
+		print "已经保存!"
+	#如果不追加日记内容, 重新写日记
+	else:
+		print "退出!"
 
-print "让我们开始日记吧!!!"
-content = raw_input(">")
+#如果日记文件名不重复,添加一条新纪录
+else:
+	print "让我们开始日记吧!!!"
+	content = raw_input(">")
 
+	txt = open(text, 'w')
 
-txt.write(content)
+	txt.write(content)
 
-txt.close
+	txt.close
 
-print "已经保存日记"
+	print "已经保存!"
 
 
 #查看历史日记
 
-print "是否要继续查看其他日记? 输入 y继续 n退出 "
+print "是否要查看其他日记? 输入 y继续 n或者\"回车\"退出 "
 
 ###Q如何命名变量,比如这个地方
 goon = raw_input("")
@@ -45,7 +64,7 @@ if goon == "y":
 
 		R.close
 	else:
-		print "您查询的日记不存在!"
+			print "您查询的日记不存在!"
 else:
 	print "退出日记"
